+++ yafsrc/YetAnotherForum.NET/pages/posts.ascx.cs
@@ -484,6 +484,16 @@ namespace YAF.Pages
         {
             if (!this.PageContext.IsGuest)
             {
+                YafContext.Current.PageElements.RegisterJsBlockStartup(
+                    "SelectedQuotingJs",
+                    JavaScriptBlocks.SelectedQuotingJs(
+                        YafBuildLink.GetLinkNotEscaped(
+                            ForumPages.postmessage,
+                            "t={0}&f={1}",
+                            this.PageContext.PageTopicID,
+                            this.PageContext.PageForumID),
+                            this.GetText("POSTS", "QUOTE_SELECTED")));
+
                 // The html code for "Favorite Topic" theme buttons.
                 string tagButtonHTML =
                     "'<a class=\"yafcssbigbutton rightItem\" href=\"javascript:addFavoriteTopic(' + res.d + ');\" onclick=\"jQuery(this).blur();\" title=\"{0}\"><span>{1}</span></a>'"